refactor: move outputs to specific modules, fix module helpers export
