+++ cli/src/modules/catalog/module.ts
@@ -1,0 +1,5 @@
+import { showTemplates } from "./output";
+
+export const CatalogModule = {
+  showTemplates,
+};
